third_party/blink/web_tests/external/wpt/fetch/api/redirect/redirect-origin.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] [POST] Redirect 307 Same origin to same origin
  assert_equals: Check origin header expected (object) null but got (string) "http://web-platform.test:8001"
[FAIL] [POST] Redirect 308 Same origin to same origin
  assert_equals: Check origin header expected (object) null but got (string) "http://web-platform.test:8001"
Harness: the test ran to completion.
